Reorganize Discord bot code with comprehensive section headers
