--- discord/REFACTORING_NOTES.md
@@ -1,105 +1,0 @@
-# Refactoring Notes
-
-## Current State
-
-The Discord bot's `index.js` file is 6944 lines with approximately 71 functions that are highly interdependent. A complete refactoring would require careful analysis of all dependencies.
-
-## What Was Successfully Refactored
-
-The following modules were successfully extracted:
-
-1. **config/constants.js** - Configuration constants
-2. **utils/permissions.js** - Permission checking functions
-3. **utils/formatting.js** - Text formatting utilities
-4. **utils/components.js** - UI component builders
-5. **services/dynamodb.js** - Database operations
-6. **services/lambda.js** - Lambda invocations
-7. **services/monitoring.js** - Status monitoring logic
-
-## What Remains in index.js
-
-The majority of the code remains in index.js because:
-
-1. **Complex interdependencies** - Many functions call each other in complex ways
-2. **Shared state** - Functions rely on the Discord client and its collections
-3. **Event handlers** - Button, modal, and select menu handlers are tightly coupled
-4. **Dashboard rendering** - The sendUnifiedDashboard function is called from many places
-
-## Recommended Approach
-
-Given the complexity, a phased approach is recommended:
-
-### Phase 1 (Current PR)
-- Extract truly independent utilities ✅
-- Create module structure ✅
-- Keep core functionality in index.js
-- Ensure 100% backward compatibility
-
-### Phase 2 (Future PR)
-- Extract button handlers one by one
-- Create proper dependency injection
-- Test each extraction thoroughly
-
-### Phase 3 (Future PR)
-- Extract modal and select menu handlers
-- Refactor dashboard rendering
-- Create proper event emitters for state changes
-
-### Phase 4 (TypeScript Migration)
-- Convert modules to TypeScript incrementally
-- Add proper types and interfaces
-- Improve error handling
-
-## Functions That Need Careful Extraction
-
-These functions have complex dependencies and should be extracted carefully:
-
-1. **handleButtonInteraction** - 200+ lines, handles all button types
-2. **sendUnifiedDashboard** - 300+ lines, complex rendering logic
-3. **handleModalSubmit** - Handles multiple modal types
-4. **refreshRegistrationStats** - Updates multiple channels
-5. **handleAdminButtonInteraction** - Complex admin logic
-
-## Current Module Structure
-
-```
-discord/
-├── index.js (still contains most functionality)
-├── config/
-│   └── constants.js ✅
-├── utils/
-│   ├── permissions.js ✅
-│   ├── formatting.js ✅
-│   ├── components.js ✅
-│   ├── channels.js ✅
-│   ├── dashboard.js (partial implementation)
-│   ├── stats.js (placeholder)
-│   └── admin.js (placeholder)
-├── services/
-│   ├── discord-client.js ✅
-│   ├── dynamodb.js ✅
-│   ├── lambda.js ✅
-│   └── monitoring.js ✅
-├── events/
-│   ├── ready.js (partial implementation)
-│   └── interactionCreate.js (partial implementation)
-└── handlers/
-    ├── buttonHandlers.js (placeholder)
-    ├── adminHandlers.js (placeholder)
-    ├── modalHandlers.js (placeholder)
-    └── selectMenuHandlers.js (placeholder)
-```
-
-## Why This Approach
-
-1. **Risk Mitigation** - Extracting everything at once risks breaking functionality
-2. **Maintainability** - Gradual refactoring allows testing at each step
-3. **Team Collaboration** - Smaller PRs are easier to review
-4. **Production Safety** - The bot continues to work throughout the refactoring
-
-## Next Steps
-
-1. Merge this PR with the basic structure
-2. Create issues for each phase of refactoring
-3. Extract functions incrementally with thorough testing
-4. Eventually achieve full modularization before TypeScript migration
