--- discord/handlers/adminHandlers.js
@@ -1,30 +1,0 @@
-const { EmbedBuilder } = require("discord.js");
-const { COLORS, EMOJIS } = require("../config/constants");
-const { hasAdminRole } = require("../utils/permissions");
-
-/**
- * Handle admin button interactions
- * @param {import('discord.js').Client} client
- * @param {import('discord.js').ButtonInteraction} interaction
- */
-async function handleAdminButton(client, interaction) {
-  // Check admin permissions
-  if (!hasAdminRole(interaction.member)) {
-    return await interaction.reply({
-      content: `${EMOJIS.ERROR} You do not have admin permissions.`,
-      ephemeral: true,
-    });
-  }
-
-  // TODO: Implement admin button handlers
-  // This would include handling admin panel buttons, user management, etc.
-  
-  await interaction.reply({
-    content: "Admin functionality will be implemented here.",
-    ephemeral: true,
-  });
-}
-
-module.exports = {
-  handleAdminButton,
-};
