--- discord/handlers/selectMenuHandlers.js
@@ -1,22 +1,0 @@
-const { EMOJIS } = require("../config/constants");
-
-/**
- * Handle select menu interactions
- * @param {import('discord.js').Client} client
- * @param {import('discord.js').StringSelectMenuInteraction} interaction
- */
-async function handleSelectMenu(client, interaction) {
-  const customId = interaction.customId;
-
-  // TODO: Implement select menu handlers
-  // This would include handling backup selection, user selection for admin, etc.
-
-  await interaction.reply({
-    content: "Select menu functionality will be implemented here.",
-    ephemeral: true,
-  });
-}
-
-module.exports = {
-  handleSelectMenu,
-};
