--- discord/utils/admin.js
@@ -1,21 +1,0 @@
-const { EmbedBuilder } = require("discord.js");
-const { COLORS, ADMIN_CATEGORY_ID } = require("../config/constants");
-const { invokeFoundryLambda } = require("../services/lambda");
-
-/**
- * Update admin status dashboard
- * @param {import('discord.js').Client} client
- */
-async function updateAdminStatus(client) {
-  try {
-    // TODO: Implement admin status update
-    // This would update an admin dashboard with system status, active users, etc.
-    console.log("Admin status update - to be implemented");
-  } catch (error) {
-    console.error("Error updating admin status:", error);
-  }
-}
-
-module.exports = {
-  updateAdminStatus,
-};
